Move onContainerSetEvent to model
